requirements.txt
accelerate
datasets==3.0.0
deepspeed==0.9.3
protobuf==3.20.3
diffusers==0.21.4
einops==0.6.1
ftfy==6.1.1
gradio==3.45.0
gradio-client==0.5.2
imageio==2.31.1
iopath==0.1.10
ipdb==0.13.13
joblib==1.3.1
matplotlib==3.7.1
mdtex2html==1.2.0
packaging==23.1
pandas==2.2.3
peft==0.15.2
Pillow==9.5.0
PyYAML==6.0.1
regex==2023.6.3
scipy
timm==0.9.2
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.3
transformers==4.46.0
sentencepiece==0.1.99
librosa==0.10.1
ffprobe==0.5
ffmpeg==1.4
fad_pytorch
tensorboard
tiktoken==0.9.0
fastai==1.0.52
pydantic==1.10.21
thinc==8.3.4
wfdb==4.3.0
blobfile==3.0.0
numpy==1.26.4
trl==0.16.1
rouge_score
bert_score
pycocoevalcap
